Extend 'reload' support
'reload' will now start a service if it isn't already running.
If no reload signal is configured the service is restarted to achieve
the reload.
Started looking at support for a reload command which would be executed
if present. The intent is to allow for servcices that may reload their
configs after receiving a command (e.g. ubus request). The actual
command isn't getting sent yet.