early start:
Added some support for starting some services early (e.g. ubusd)
The serviced command line accepts a -d <filespec> option that indicates
the filespec of json files to be processed early on. ubusd is a good
example of a service that needs to be started early as it is needed
by serviced to be able to send events etc.